Change structure and refactoring
