+++ ORM_TST/Constants.cs
@@ -1,0 +1,6 @@
+public class Constants{
+
+    public const string TABLE_NAME = "CUSTOMER";
+    public const string COLUMN_NAME = "CLIENT_ID";
+    public const string PROPERTY_NAME = "ClientID";
+}
